Exclude frames before main
I previously avoided this logic as this was not affecting Rust profile.

Now though, I'm going to change the profiler to unwind through frame pointer when it's running on mac.
When this happen though, we might incur in frames occurring before the main function as it's currently
happening in iOS.

This will fix it.