feat(functions): Collapse the call tree to reduce noise
Returning the entire call tree results in a lot of noise. This change aims to
reduce the noise by targeting the case where a frame in the call tree has
exactly 1 child, and that child's start/end times matches the parent. In this
case, showing both the parent and the child is meaning less as they will always
have the same durations and convey roughly the same information. So, we opt to
show the inner most application frame, or the inner most frame if there are no
application frames.