feat(functions): Remove single sampled frames when collapsing calltree
Single sampled frames are not useful information to show in suspect functions.
This filters them out of the call tree so that they are not inserted into snuba
and not shown in the UI. This also potentially has the effect of reducing the
size of the calltree written to kafka and reduce memory usage.