fix(profiling): Only update frame package for native platform
Package should only be used for native platforms like cocoa and rust, make sure
it's only update for those. This only updates it on the frame and not the
nodetree because the nodetree is serialized for the suspect functions feature
which coalesces module and package together still so that is left untouched.
This also ensures that the frame fingerprint does not change.