feat(suspect-functions): New suspect functions generation
This changes the way we extract functions for the suspect functions feature.
Previously, we used the frame duration, this change switches us over to use self
times. For system frames, the self time is how it is normally defined, but for
application frames, the self time is the frame duration minus the frame
durations of any child application frames. This means that system frames do not
affect the self time of application frames.