ref(legacy): Make legacy profile the promoted type
By making the legacy profile the promoted type, this allows us the make the
legacy android + ios profile types the top level type, giving us access to the
various metadata fields when we work with the raw data.

This also has the added benefit of decoupling the legacy ios and android
profiles so the legacy ios profile can be easily removed when we drop support
for it.