reuse one http client across different requests
We should always reuse an http client across different requests as
they're thread-safe and meant to be reused for better performance
(connection pooling).

This will avoid creating a new one (and relative http transport) for
each and every requests we make to send metrics.

Not doing this might lead to temporary port exhaustion and degradation
of perfomances.