perf(flamegraph): Move calltree generation into readjob
Currently we read from gcs in a pool of workers but the calltree generation
happens sequentially in the original request handler. This becomes a bottle neck
as we try to generate calltrees for more profiles. So move the calltree
generation into the goroutines so it happens in parallel and saturates the CPU.