refactor sliceCallTree so that it won't change the original callTRee
Currently while sliceCallTree returns a newer sliced callTree, it also
changes the internals of the original one.

This fixes it by creating a shallow copy of the nodes whenever we're
going to modify them due to slicing.

This means we can safely call sliceCallTree multiple times on the
original callTree.