chore(flamegraph): Lower number of concurrent reads
The new flamegraph endpoint is very likely the suspect of the OOMs we've been
seeing. Upon examining the old flamegraph endpoint, we used to spawn ~25 go
routines per request but at a lower throughput. So having 100 concurrent workers
may be the cause of the OOMs as we try to read too many profiles at once.