fix android callTree timestamps and make them consistent with the sample format
Android and sample format share the same CallTree struct.
While the sample format (v2) set the ts of the nodes to as absolute ts,
the android one is setting a relative one.
This break the callTree slice.

An optional SdkStartTime has been added to optionally convert the ts to
absolute ones.

For legacy android profiles we won't set it, hence it'll keep the
relative ones.