fix(docker): Add entry point script, fix permissions
When we switched to using a non-root user as default, it broke our default storage location which was under `/var/lib`. That place is only writeable by root, and for good reason.

This PR refactors how we manage the storage options by adding an entry point script which also takes care of reowning the storage folder, as this would be needed for the non-root user transition.